Move the fixed JSONRPC class to the rpc package
