+++ rpc/__init__.py
@@ -1,3 +1,4 @@
 from .util import authenticate
+from .fixed_jsonrpc import FixedJSONRPC
 
-__all__ = ['authenticate']
+__all__ = ['authenticate', 'FixedJSONRPC']
